# backend/src/NuamExchange.Api/Controllers/ClassificationsController.cs
using Microsoft.AspNetCore.Mvc; using NuamExchange.Application.Classifications; using NuamExchange.Application.Common;
namespace NuamExchange.Api.Controllers;
[ApiController][Route("api/v1/classifications")]
public sealed class ClassificationsController(IClassificationService service):ControllerBase
{ [HttpGet][ProducesResponseType(typeof(PaginatedResponse<ClassificationDto>),200)] public async Task<ActionResult<PaginatedResponse<ClassificationDto>>> List([FromQuery] ClassificationListRequest request,CancellationToken ct)=>Ok(await service.ListAsync(request,ct));
 [HttpGet("catalogs")][ProducesResponseType(typeof(ClassificationCatalogsDto),200)] public async Task<ActionResult<ClassificationCatalogsDto>> Catalogs(CancellationToken ct)=>Ok(await service.GetCatalogsAsync(ct));
 [HttpGet("{id}")][ProducesResponseType(typeof(ClassificationDto),200)][ProducesResponseType(404)] public async Task<ActionResult<ClassificationDto>> Get(string id,CancellationToken ct){var dto=await service.GetByIdAsync(id,ct);SetEtag(dto.RowVersion);return Ok(dto);} 
 [HttpPost][ProducesResponseType(typeof(ClassificationWriteResponse),201)][ProducesResponseType(400)][ProducesResponseType(409)] public async Task<ActionResult<ClassificationWriteResponse>> Create([FromBody] CreateClassificationRequest request,CancellationToken ct){var r=await service.CreateAsync(request,ct);SetEtag(r.Classification.RowVersion);return CreatedAtAction(nameof(Get),new{id=r.Classification.Id},r);} 
 [HttpPut("{id}")][ProducesResponseType(typeof(ClassificationWriteResponse),200)] public async Task<ActionResult<ClassificationWriteResponse>> Update(string id,[FromBody] UpdateClassificationRequest request,[FromHeader(Name="If-Match")] string? ifMatch,CancellationToken ct){var r=await service.UpdateAsync(id,request,ifMatch,ct);SetEtag(r.Classification.RowVersion);return Ok(r);} 
 [HttpPost("{id}/copy")][ProducesResponseType(typeof(ClassificationWriteResponse),201)] public async Task<ActionResult<ClassificationWriteResponse>> Copy(string id,[FromBody] CopyClassificationRequest request,CancellationToken ct){var r=await service.CopyAsync(id,request,ct);SetEtag(r.Classification.RowVersion);return CreatedAtAction(nameof(Get),new{id=r.Classification.Id},r);} 
 [HttpDelete("{id}")][ProducesResponseType(204)] public async Task<IActionResult> Delete(string id,[FromHeader(Name="If-Match")] string? ifMatch,CancellationToken ct){await service.DeleteAsync(id,ifMatch,ct);return NoContent();} void SetEtag(string rv)=>Response.Headers.ETag=$"\"{rv}\"";}