# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: NUAM Exchange API
  version: 0.1.0
  description: Contrato provisional para integración futura con ASP.NET Core Web API .NET 8.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /classifications:
    get:
      tags: [Classifications]
      operationId: listClassifications
      parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' }, { name: search, in: query, schema: { type: string } }, { name: market, in: query, schema: { type: string } }, { name: source, in: query, schema: { type: string } }, { name: fiscalYear, in: query, schema: { type: integer } }, { name: status, in: query, schema: { type: string } }, { name: sortBy, in: query, schema: { type: string } }, { name: sortDirection, in: query, schema: { $ref: '#/components/schemas/SortDirection' } } ]
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedClassifications' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '500': { $ref: '#/components/responses/ServerError' } }
    post: { tags: [Classifications], operationId: createClassification, requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateClassificationRequest' } } } }, responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/ClassificationWriteResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' }, '500': { $ref: '#/components/responses/ServerError' } } }
  /classifications/catalogs:
    get: { tags: [Classifications], operationId: getClassificationCatalogs, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ClassificationCatalogs' } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '500': { $ref: '#/components/responses/ServerError' } } }
  /classifications/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get: { tags: [Classifications], operationId: getClassificationById, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Classification' } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '404': { $ref: '#/components/responses/NotFound' }, '500': { $ref: '#/components/responses/ServerError' } } }
    put: { tags: [Classifications], operationId: updateClassification, parameters: [ { $ref: '#/components/parameters/IfMatch' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateClassificationRequest' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ClassificationWriteResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '409': { $ref: '#/components/responses/Conflict' }, '500': { $ref: '#/components/responses/ServerError' } } }
    delete: { tags: [Classifications], operationId: deleteClassification, parameters: [ { $ref: '#/components/parameters/IfMatch' } ], responses: { '204': { description: No Content }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' }, '500': { $ref: '#/components/responses/ServerError' } } }
  /classifications/{id}/copy:
    post: { tags: [Classifications], operationId: copyClassification, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/CopyClassificationRequest' } } } }, responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/ClassificationWriteResponse' } } } }, '404': { $ref: '#/components/responses/NotFound' }, '500': { $ref: '#/components/responses/ServerError' } } }
...
components:
  securitySchemes: { bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT } }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string } }
    IfMatch: { name: If-Match, in: header, required: false, schema: { type: string } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, schema: { type: integer, enum: [5, 10, 20], default: 10 } }
  responses:
    BadRequest: { description: Bad Request, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
    Unauthorized: { description: Unauthorized, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
    Forbidden: { description: Forbidden, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
    NotFound: { description: Not Found, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
    Conflict: { description: Conflict, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
    ServerError: { description: Server Error, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
  schemas:
    SortDirection: { type: string, enum: [asc, desc] }
    ProblemDetails: { type: object, required: [title, status], properties: { type: { type: string }, title: { type: string }, status: { type: integer }, detail: { type: string }, instance: { type: string }, traceId: { type: string }, errors: { type: object, additionalProperties: { type: array, items: { type: string } } } } }
...
    Classification: { type: object, required: [id, market, source, fiscalYear, instrument, paymentDate, description, eventSequence, updateFactor, amount, status], properties: { id: { type: string }, market: { type: string }, source: { type: string }, fiscalYear: { type: integer }, instrument: { type: string }, paymentDate: { type: string, format: date }, description: { type: string }, eventSequence: { type: string }, updateFactor: { type: number, format: double }, amount: { type: number, format: double }, status: { type: string }, rowVersion: { type: string, format: byte } } }
    CreateClassificationRequest: { allOf: [ { $ref: '#/components/schemas/Classification' } ] }
    UpdateClassificationRequest: { allOf: [ { $ref: '#/components/schemas/Classification' } ] }
    CopyClassificationRequest: { type: object, properties: { paymentDate: { type: string, format: date }, description: { type: string }, eventSequence: { type: string } } }
    ClassificationCatalogs: { type: object, properties: { markets: { type: array, items: { type: string } }, sources: { type: array, items: { type: string } }, fiscalYears: { type: array, items: { type: integer } }, statuses: { type: array, items: { type: string } } } }
    PaginatedClassifications: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Classification' } }, page: { type: integer }, pageSize: { type: integer }, totalItems: { type: integer }, totalPages: { type: integer } } }
    ClassificationWriteResponse: { type: object, properties: { classification: { $ref: '#/components/schemas/Classification' }, message: { type: string } } }
...